+++ public/assets/json/categorias/colchones/sub-categorias/rucay.json
@@ -44,6 +44,7 @@
                     "ruta": "/productos/colchones/rucay/king/"
                 }
             ],
+            "tipo-de-envio": "Gratis",
             "detalles-del-producto":[
                 {
                     "Sku": "K1152CN10",
